Phase 3 (Bug fixes)

+++ app/Models/Agreement.php
@@ -6,7 +6,7 @@ use Carbon\Carbon;
 use Illuminate\Support\Facades\Hash;
 
 /**
- * Class Agreement
+ * Model for agreements table.
  *
  * @package App\Models
  * @property int id
@@ -61,7 +61,9 @@ class Agreement extends Model
     {
         $user = $this->company->user;
 
-        return $user->delete();
+        if ($user != null) {
+            return $user->delete();
+        }
     }
 
     public static function expiredToday()
